Bug 2027694 - Remove manual DomainMatchingMode cast.  a=diannaS DONTBUILD
Original Revision: https://phabricator.services.mozilla.com/D292859

Differential Revision: https://phabricator.services.mozilla.com/D293531

+++ dom/storage/SessionStorageManager.cpp
@@ -653,8 +653,7 @@ void SessionStorageManager::ClearStorages(
       return;
     }
 
-    mActor->SendClearStorages(aPattern, nsCString(aOriginScope),
-                              static_cast<uint32_t>(aMode));
+    mActor->SendClearStorages(aPattern, nsCString(aOriginScope), aMode);
   }
 
   ClearStoragesInternal(aPattern, aOriginScope, aMode);
